[FE] Refactor: GerneCardList scrollList 포함하지 않도록 수정
- GenreCardList index.tsx 파일 추가 기존 파일 삭제
- ChartCardList  index.tsx 파일 추가 기존 파일 삭제
